move helper functions to utils events js
