+++ src/view/header-trip-info.js
@@ -4,7 +4,7 @@ export const createHeaderTripInfoTemplate = ({route, duration, cost}) => {
   return (
     `<section class="trip-main__trip-info  trip-info">
       <div class="trip-info__main">
-        <h1 class="trip-info__title">${route}</h1>
+        <h1 class="trip-info__title">${route.join(` &mdash; `)}</h1>
 
         <p class="trip-info__dates">${duration}</p>
       </div>
